Add RichTextBuilder
