+++ druid/src/text/mod.rs
@@ -23,6 +23,7 @@ pub mod format;
 mod layout;
 pub mod movement;
 pub mod selection;
+mod rich_text;
 mod storage;
 mod text_input;
 
@@ -35,4 +36,5 @@ pub use self::movement::{movement, Movement};
 pub use self::selection::Selection;
 pub use self::text_input::{BasicTextInput, EditAction, MouseAction, TextInput};
 pub use editor::Editor;
-pub use storage::{ArcStr, RichText, TextStorage};
+pub use storage::{ArcStr, TextStorage};
+pub use rich_text::{RichText, RichTextBuilder, AttributesAdder};
