+++ druid/src/text/rich_text.rs
@@ -1,0 +1,147 @@
+// Copyright 2018 The Druid Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//! Rich text with style spans.
+
+use std::ops::{Range, RangeBounds};
+use std::sync::Arc;
+
+use super::{Attribute, AttributeSpans, TextStorage};
+use crate::piet::{util, PietTextLayoutBuilder, TextLayoutBuilder, TextStorage as PietTextStorage};
+use crate::{ArcStr, Data, Env};
+
+/// Text with optional style spans.
+#[derive(Debug, Clone, Data)]
+pub struct RichText {
+    buffer: ArcStr,
+    attrs: Arc<AttributeSpans>,
+}
+
+impl RichText {
+    /// Create a new `RichText` object with the provided text.
+    pub fn new(buffer: ArcStr) -> Self {
+        RichText::new_with_attributes(buffer, Default::default())
+    }
+
+    /// Create a new `RichText`, providing explicit attributes.
+    pub fn new_with_attributes(buffer: ArcStr, attributes: AttributeSpans) -> Self {
+        RichText {
+            buffer,
+            attrs: Arc::new(attributes),
+        }
+    }
+
+    /// Builder-style method for adding an [`Attribute`] to a range of text.
+    ///
+    /// [`Attribute`]: enum.Attribute.html
+    pub fn with_attribute(mut self, range: impl RangeBounds<usize>, attr: Attribute) -> Self {
+        self.add_attribute(range, attr);
+        self
+    }
+
+    /// The length of the buffer, in utf8 code units.
+    pub fn len(&self) -> usize {
+        self.buffer.len()
+    }
+
+    /// Returns `true` if the underlying buffer is empty.
+    pub fn is_empty(&self) -> bool {
+        self.buffer.is_empty()
+    }
+
+    /// Add an [`Attribute`] to the provided range of text.
+    ///
+    /// [`Attribute`]: enum.Attribute.html
+    pub fn add_attribute(&mut self, range: impl RangeBounds<usize>, attr: Attribute) {
+        let range = util::resolve_range(range, self.buffer.len());
+        Arc::make_mut(&mut self.attrs).add(range, attr);
+    }
+}
+
+impl PietTextStorage for RichText {
+    fn as_str(&self) -> &str {
+        self.buffer.as_str()
+    }
+}
+
+impl TextStorage for RichText {
+    fn add_attributes(
+        &self,
+        mut builder: PietTextLayoutBuilder,
+        env: &Env,
+    ) -> PietTextLayoutBuilder {
+        for (range, attr) in self.attrs.to_piet_attrs(env) {
+            builder = builder.range_attribute(range, attr);
+        }
+        builder
+    }
+}
+
+/// A builder for [`RichText`].
+///
+/// # Example
+/// ```
+/// let mut rich_text = RichTextBuilder::new();
+/// rich_text.push("Hello World").add(Attribute::Underline(true));
+/// let rich_text = rich_text.build();
+/// ```
+///
+/// [`RichText`]: RichText
+#[derive(Debug, Default)]
+pub struct RichTextBuilder {
+    buffer: String,
+    attrs: AttributeSpans,
+}
+
+impl RichTextBuilder {
+    /// Create a new `RichTextBuilder`.
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    /// Add string to the end of text and `AttributesBuilder` for the added string.
+    pub fn push(&mut self, string: &str) -> AttributesAdder {
+        let range = self.buffer.len()..(self.buffer.len() + string.len());
+        self.buffer.push_str(string);
+        self.range(range)
+    }
+
+    /// Get the `AttributesBuilder` for the range.
+    pub fn range(&mut self, range: Range<usize>) -> AttributesAdder {
+        AttributesAdder {
+            rich_text_builder: self,
+            range,
+        }
+    }
+
+    /// Build the `RichText`.
+    pub fn build(self) -> RichText {
+        RichText::new_with_attributes(self.buffer.into(), self.attrs)
+    }
+}
+/// Adds Attributes to the text.
+///
+/// See also: [`RichTextBuilder`](RichTextBuilder)
+pub struct AttributesAdder<'a> {
+    rich_text_builder: &'a mut RichTextBuilder,
+    range: Range<usize>,
+}
+
+impl AttributesAdder<'_> {
+    /// Add the given attribute.
+    pub fn add(&mut self, attr: Attribute) -> &mut Self {
+        self.rich_text_builder.attrs.add(self.range.clone(), attr);
+        self
+    }
+}
